console: Refactor connect in owners select
